Theoretical re-configuration of the canvas
But when initially set to opaque, the premultiply option goes alhpa
to black instead of the background.

The gpu might need to be rebuilt? or there is a `unconfigure` for
canvas that may need to be done, but it's experimental. I'm going
to look at other resources to see how they are doing it.