Reorg - lib build for vite; experiments is out of source
I'd like to convert the experiments directory and entry point into
github pages, and I also am feeling confident enough about the
utility of this lib to start building/exporting a lib.

So, this separates the experiments from the library. It creates a
separate vite config for generating the library. Type export not
yet happening, sadly.