Changes Model interface to work with collections of buffers
Adds some quad and point models and sketches out how to create a
vertex and index buffer for a single quad (the background). But
doesn't actually integrate it. Spiking that out provided a different
interface for models so that they could generate multiple buffers.
That meant changing the methods to plurals, but also moving towards
collection which are easier in some ways (never null) and harder
in others (maybe there is a missing Buffer class that should write
to the gpu and provide a description?).