Reworks models again!
The vertex buffer has to write to the pipeline in a specific way.
Same with indexes etc. The pipeline just needs to ask questions
directly to the buffers. So, model are responsible for creating
their own buffers, and the pipeline interacts with the buffers
to write to the device etc.

This is still only working with the uniform color model, but
sketched out he vertex and index buffers.