Separates poylgon to storage concerns from mutation
A bit.

This isn't integrated, but the algorithm is tested and works. There
is a class now PolygonStorage that manages the conversion of a bunch
of polygons into vertices (unique) and indices (for triangle rendering).
It uses a subclass for ease of programming.

I'm still envisioning a Subpoly class that will read data and also allow
modification.

Adding and deleting polygons is still a mystery. Also, converting to
buffers and reconstructing those buffers after an addition or deletion
is still a mystery to unfold.